+++ changelog.md
@@ -1,5 +1,7 @@
-* new structure of project (prepare/parser/builder)
+* new structure of project (prepare/parser/builder/wrapper)
 * new parser(less stronger then earlier) based on htmlparser (why is it important, question of performance)
 * tests
 * throw of parsing exception
 * comments in itemplate functions
+
+todo extended mode for attributes